employer/recruiter_app_fixed_v2/requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.5
flask-login==0.6.2
flask-wtf==1.1.1
email-validator==1.3.1
flask-babel==2.0.0
python-dotenv==1.0.0
